ENH: geometric_slerp extrapolation
* Fixes gh-23400.

* Add support and regression tests for extrapolation behavior
in `geometric_slerp()`, and update the documentation for
the interpolation parameter accordingly.

* The `Raises` section of the `geometric_slerp()` docstring
was also updated to remove a spurious claim about antipodes--we
now actually allow i.e., North and South pole points despite
the infinite number of paths between them on the sphere, and
enforce this behavior in our regression tests (a warning, but not error,
is issued). That is not directly related to the changes here, but
the documented innacurracy of behavior would otherwise make this
harder to reason about.

[ci skip] [skip ci]