Fix documentation bug in Rotation.as_davenport()
Fix the documentation giving the wrong range of possible values for the second axis, when lambda is negative.

If lambda is negative, the code will flip the second axis, calculate it with the now positive value of lambda and flip the sign of the second calculated angle. Therefore the range of possible value for negative lambda the range of possible values for positive lambda, but mirrored at zero.