BUG: stats: avoid infinite loop in low_0_bit
There are two bugs in this code. First, (1 << i) done as a 32-bit
operation. This means that if the first 0 bit is at the 33rd bit,
then low_0_bit cannot find it. This means that the 64 bit
specialization of this code is incorrect.

Second, if x is all 1s, then finding the first zero bit will never
finish on x86,as << wraps around if a shift larger than the number of
bits in the input is requested. (For example, a shift of 33 on a uint32
becomes a shift of 1.) Formally, this is UB.