DOC:signal._signaltools.detrend: Rename overwrite_data.
The argument `overwrite_data=True` made it appear as if overwriting was
guaranteed. However, overwriting only happens when ``type == linear``
and ``dtype is in "dtDF"`` and array reshaping returns a view.
I changed the argument name to `allow_overwrite` and updated the
docstring to indicate overwriting only happens under specific
conditions.