ENH/BENCH: Remove 1-norm computation from SuperLU object creation.
This commit effectively reverts to b3c80ed, but includes intermediate
changes based on reviewer comments, so it is not actually a "revert"
commit. Also includes minor documentation updates.

We now compute the 1-norm of the input matrix A using
``sparse.linalg.norm`` instead of internally in the SuperLU object.
Thus, the norm will only be computed when `cond1est` is called, not for
every SuperLU object that is created.

This commit also includse a benchmark test for comparison with the
internal C norm computation version.