BUG: signal: Control output dtypes in `abcd_normalize` by adding new keyword parameter.
With this commit `signal.abcd_normalize` gains the capability of controlling the dtype of the returned arrays through the new keyword parameter `cast_to`. The new default behavior is to make the output dtype float64 / complex128 (depending on input dtypes). Alternatively, either a specific dtype can be provided or type conversion can be turned off.

* `signal.ss2tf` now converts parameters to floating due to relying on `abcd_normalize`.
* The initializer of class `Statespace` gained the keyword `cast_to`, which is passed to `abcd_normalize` for normalizing matrix parameters. Hence, those are converted to floating by default.
* Unit tests were added to ensure 100% coverage.
* Marginal docstr improvements.
* Supersedes PR 13235 from 2020: The `abcd_normalize` code diverged too much for making sensible adaptions in that PR.
* Closes issue 13125, closes issue 18982.