BUG: stats: Fix incorrect handling of the trimmed mean in `levene` and `fligner`. (#23557)
Change the implementation of the Levene test in the case where the selected
center is the trimmed mean, in order to trim the data within each sample
only when computing the center, instead of trimming the data for the whole test
as was the case before.

The corresponding unit test is changed accordingly, and the results are
tested against values obtained with R's leveneTest method from the car package.
An additional test is added for the case where the selected center is the mean,
also tested against the same R function.

A similar change is made in the implementation of the Fligner test.