TST: Create _CompressedMixin class for test_setdiag* tests.
This test creates a new mixin class for testing the
`_cs_matrix._setdiag` function branches from gh-19962. As-is, the
`_setdiag` function has a bug in the convert to COO branch that does not
sort the indices if `self.has_sorted_indices` is set. The test is marked
with an `xfail`.