BUG/MNT: Update condition for converting to COO format in _setdiag.
This commit changes the condition for converting from CSC/CSR format to
COO format in `_cs_matrix._setdiag.` The previous condition was:

  self.nnz - mask.sum() < self.nnz * 0.001

where `mask` was the logical array that was True for existing elements
on the k-th diagonal. This condition did not have the intended effect
when there were non-zeros *not* on the diagonal being set. The new
condition is:

  N_new < self.nnz * 0.001

where `N_new` is the number of non-existing elements on the k-th
diagonal.

This commit refactors the `mask` array to be named explicitly
`is_existing` because it represents the logical indexing of existing
elements. Similarly, `is_new = ~is_existing`.