PoC/Experiment: Type based dispatching for array-api (via spatch)
Old message below, the latest version uses a decorator to make the
code layout clearer.
It also uses the `['argname', 'argname2']` syntax as it is clearer
and shorter.  That may be slower, but if a problem it can also be
optimized in spatch (although only with C probably) and that seems
almost more worthwhile to me.

In the current state, assumes forwarding strings is actually OK
although it may not be in practice.  Since strings are never an
interesting type to dispatch.

---

This is very much a PoC/experiment, for example the `xp=` won't work.

The backend lists the types we need. Things that need to be figured out
for example:
* We need a way to get the types from `xp` (maybe via the namespace info)
  or redefine the type to be the same thing as `xp` (but using types seems
  nice for a start).
  (within spatch, I would use `type=cupy.ndarray` right now instead, but
  `xp` makes perfect sense to me as an API here as well)
* I would need a bit of `__getattr__` magic to drop the `xp=` or add
  machinery to spatch to do so.
  (I don't want to add too much such things to `spatch` because I prefer
  the backend author to do extra work over forcing adopting libs to do it.)

Just tried with JAX locally, I am sure much won't work.

Spatch of course could expose a context manager to do more, but we don't need
to do that.

One thing we *can* do, is to create a generic `ArrayAPI` backend/version (not
the NumPy fallback!) so that then, say `torch` could add a backend to replace
the generic version with a better one where desired.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>