MAINT: clean up some outdated CPython C API idioms
Motivated by the discussion in scipy#19878. With these changes,
all Cython code can be built against the Limited API. That just requires adding
`-DPy_LIMITED_API=0x030C0000` as a flag to `cython_c_args` in scipy/meson.build

Note that using the Limited API isn't all that useful yet, since it's blocked
by lack of support in dependencies other than Cython. But it's a start, and makes
it easier to do benchmarking of the functionality implemented in Cython.