ENH: Add handling of double type for _gen_harmonic, _normalized_gen_harmonic.
The ufuncs _gen_harmonic and _normalized_gen_harmonic now have native
loops for type double for the parameters that are normally integers.

This allows removal of the integer cast in the methods of stats.zipfian.

To avoid the possibility of infinite loops and other numerical problems,
the `n` parameter of stats.zipfian must now be less than 2**53.

Closes gh-23678.