BUG: Fixes #23644: sparse: csc_array.setdiag() makes indices unsorted (#23676)
* TST: Create _CompressedMixin class for test_setdiag* tests.

This test creates a new mixin class for testing the
`_cs_matrix._setdiag` function branches from gh-19962. As-is, the
`_setdiag` function has a bug in the convert to COO branch that does not
sort the indices if `self.has_sorted_indices` is set. The test is marked
with an `xfail`.

* TST: Remove test_setdiag_cs[cr].

The existing `test_setdiag_cs[cr]` methods are now encompassed by the
`_CompressedMixin.test_setdiag_noconvert` method. Remove them to avoid
repeating logic and redundant tests.

* BUG/MNT: Update condition for converting to COO format in _setdiag.

This commit changes the condition for converting from CSC/CSR format to
COO format in `_cs_matrix._setdiag.` The previous condition was:

  self.nnz - mask.sum() < self.nnz * 0.001

where `mask` was the logical array that was True for existing elements
on the k-th diagonal. This condition did not have the intended effect
when there were non-zeros *not* on the diagonal being set. The new
condition is:

  N_new < self.nnz * 0.001

where `N_new` is the number of non-existing elements on the k-th
diagonal.

This commit refactors the `mask` array to be named explicitly
`is_existing` because it represents the logical indexing of existing
elements. Similarly, `is_new = ~is_existing`.

* BUG: Fixes gh-23644 by sorting COO indices in _cs_matrix._setdiag.

This commit checks whether `has_sorted_indices` is set, and sorts the
newly-created diagonal indices accordingly.

* TST: Remove xfail from test_setdiag_cooconvert.

The test now passes since the bug is fixed, so remove the xfail.

* BUG: Fix potential bug in evaluating has_sorted_indices in _setdiag.

This commit is in response to a reviewer comment, noting that
`has_sorted_indices` should be cached before changes are made to the
matrix, like in `_insert_many`.

* ENH: Move is_new allocation to relevant branch.

This commit is in response to a reviewer suggestion to move the
allocation of `is_new` in `_cs_matrix._setdiag` to the relevant logic
branch. It delays allocation until necessary.

* ENH: Use exsiting memory to compute logical not in _cs_matrix._setdiag.

This commit is in response to a reviewer comment. It uses
alraedy-allocated memory to compute a logical not array and reduce RAM
footprint.