ENH: special: add handling of double type for `_gen_harmonic`, `_normalized_gen_harmonic` (#23863)
* ENH: Add handling of double type for _gen_harmonic, _normalized_gen_harmonic.

The ufuncs _gen_harmonic and _normalized_gen_harmonic now have native
loops for type double for the parameters that are normally integers.

This allows removal of the integer cast in the methods of stats.zipfian.

To avoid the possibility of infinite loops and other numerical problems,
the `n` parameter of stats.zipfian must now be less than 2**53.

Closes gh-23678.

* FIXUP: Improve handling of n that is too big. Also note max n in zipfian class docstring.

* FIXUP: fix dtype used for checking n in _argcheck()

* ENH: special: Add native handling of int32 by _gen_harmonic and _normalized_gen_harmonic