BENCH: interpolate: fix the RGI subclassing benchmark (#23954)
https://github.com/scipy/scipy/pull/23759 changed RGI.values attribute into a
read-only property, and this broke the RegularGridInterpolatorSubclass benchmark.

The benchmark itself was added in https://github.com/scipy/scipy/pull/17230,
for (somewhat unconventional) downstream usage, with a caveat: this downstream
usage is reaching into the RGI internals, which is subject to change in future
SciPy versions: https://github.com/scipy/scipy/pull/17230#issuecomment-1288041388
And this "future scipy version" has arrived, in scipy==1.17.0

I tried adding a setter for the RGI.values property, but this does not seem
to help unbreaking the benchmark.

closes gh-23953