BUG: optimize: allow user callbacks to use lazily defined types
In Python 3.14, functions may use types that are not defined in the
enclosing scope. If any other function uses `inspect.signature()`
on this function, this will cause an error. Fix this by using
FORWARDREF to specify that we don't want to evaluate the annotations
while inspecting the signature. This will instead return a ForwardRef
object from the inspection.