ENH: Return spsolve solution matching shape and sparsity to RHS.
This commit changes the behavior of scipy.sparse.linalg.spsolve to match
that of np.linalg.solve and scipy.linalg.solve. In particular, when
given an RHS b shape (N, 1), the output is now (N, 1), and either sparse
or dense depending on the type of the input. Previously, the output was
flattened to (N,), which was inconsistent with the dense solve
functions.