MAINT: signal: tighten the workaround for object arrays (#23668)
* MAINT: signal: tighten the workaround for object arrays

Now that the deprecation of object arrays in lfilter et al has expired,
tighten the try-except stanza in `_support_alternative_backends`.
Turns out it was masking several small errors elsewhere, so fix them.

The only remaining instance of object array is in `tf2ss`, and it is
explicitly tested for. We might want to deprecate these, separately.

* TST: signal: xfail failing tests