TST/MAINT: optimize: thread-safety fixes for COBYQA and tests (#24041)
* TST: optimize: remove thread_unsafe entries lacking reason

* TST: optimize.basinhopping: Avoid sharing Storage object

* TST: optimize: test COBYQA in parallel

* MAINT: optimize: update COBYQA

* TST: optimize: bump DifferentialEvolutionSolver tolerance

This test looks like a tautology, but it's actually not. If local solution
polishing is enabled, then the solution will be polished in unscaled
coordinates. Then, the solution will be re-added to the population using
scaled coordinates. This means that result.x is returned directly from
the local solver, while DifferentialEvolutionSolver.x goes through
unscaling, then scaling. This can cause the result to differ by ~1e-17.