ENH: Raise LinAlgError for singular matrix in spsolve.
This commit changes the behavior of spsolve with a dense `b` input.
Previously, it would warn with MatrixRankWarning and return a vector of
np.nan values. For a sparse `b` input, spsolve would raise. Now, it
raises on either sparse or dense input.

This commit removes the MatrixRankWarning class, as it is no longer
needed.

This commit changes the type of error raised for a sparse `b` input.
SuperLU raises a RuntimeError when factorizing a singular matrix.
Previously, spsolve just raised that error. Now, it raises a LinAlgError
from that error.

The error behavior for a singular matrix is now consistent across all
input types and use_umfpack=True/False.