ENH: Remove _superlu.gssv call and only allow CSC format in spsolve.
This commit is the result of performance testing showing that gssv calls
are equivalent to splu(A).solve(b) for CSC input, but substantially
slower for large matrices and many RHS for CSR input. It is faster to
just convert the input to CSC format and call splu(A).solve(b) for all
but small matrices (N ~ 400) and/or small numbers of RHS (K < 10).