ENH: Refactor spsolve(B.T, e) -> splu(B).solve(e, trans='T').
This commit is in response to the removal of the _superlu.gssv call in
sparse.linalg.spsolve, which requires spsolve to only accept CSC inputs.

The B matrix is in CSC format, so its transpose is in CSR format, which
is not supported by spsolve. It is more efficient to create the
SuperLU factorization of the CSC format, then call the solve method with
`trans='T'` to handle the "transpose" internally at the C level.

This change forces use of SuperLU instead of the more generic spsolve,
but a future commit will support the `trans` argument to spsolve.